Actually fix bug
The previous commit did not correctly fix the bug (an additional
copy of the list of well contents appearing each time the well-
assignment modal was opened).

The root cause was that on each click on the link, an additional
callabck was being bound to the ``shown.bs.modal`` event.