Fix bug in code generation
When selecting a node's parent, was selecting them in the order in
which they were listed in the nodes array, rather than in the order
in which they were listed in node.parentIds. This meant that details
attached to edges (volumes, mixing, etc.) would sometimes be applied
to the wrong transfer.