Fix bug that could create spurious edges when nodes selected
If a protocol contained two isomophic connected components, these
can swap positions when a node is selected. This meant that the
mouseup and mousedown would different nodes (one on each component).
This meant that for some geaphs, clicking on a node would create an
edge.

To prevent this, I stop tick() from updating node positions when
mousedown_node is set. It owuld be better to swap the components
swapping positions, but I'm not sure how to do this.